+++ litebird_sim/detectors.py
@@ -12,6 +12,10 @@ from .quaternions import quat_rotation_y, quat_rotation_z, quat_left_multiply
 from .bandpasses import BandPassInfo
 
 
+def url_to_uuid(url: str) -> UUID:
+    return UUID([x for x in url.split("/") if x != ""][-1])
+
+
 @dataclass
 class DetectorInfo:
     """A class wrapping the basic information about a detector.
@@ -50,6 +54,7 @@ class DetectorInfo:
 
         - sampling_rate_hz (float): The sampling rate of the ADC
              associated with this detector. The default is 0.0
+
         - fwhm_arcmin: (float): The Full Width Half Maximum of the
              radiation pattern associated with the detector, in
              arcminutes. The default is 0.0
@@ -59,10 +64,14 @@ class DetectorInfo:
 
         - net_ukrts (float): The noise equivalent temperature of the
              signal produced by the detector in nominal conditions,
-             expressed in μK/√s.. The default is 0.0
+             expressed in μK/√s. This is the noise per sample to be
+             used when adding noise to the timelines. The default is 0.0
 
         - pol_sensitivity_ukarcmin (float): The detector sensitivity
-            in microK_arcmin. The default is 0.0
+            in microK_arcmin. This value considers the effect of cosmic ray loss,
+            repointing maneuvers, etc., and other issues that cause loss of
+            integration time. Therefore, it should **not** be used with the
+            functions that add noise to the timelines. The default is 0.0
 
         - bandcenter_ghz (float): The center frequency of the
              detector, in GHz. The default is 0.0
@@ -263,7 +272,12 @@ class FreqChannelInfo:
         # proper UUIDs
         for det_idx, det_obj in enumerate(self.detector_objs):
             if not isinstance(det_obj, UUID):
-                self.detector_objs[det_idx] = UUID(det_obj)
+                if "/" in det_obj:
+                    cur_uuid = url_to_uuid(det_obj)
+                else:
+                    cur_uuid = UUID(det_obj)
+
+                self.detector_objs[det_idx] = cur_uuid
 
     @staticmethod
     def from_dict(dictionary: Dict[str, Any]):
@@ -322,7 +336,12 @@ class InstrumentInfo:
 
         for det_idx, det_obj in enumerate(self.channel_objs):
             if not isinstance(det_obj, UUID):
-                self.channel_objs[det_idx] = UUID(det_obj)
+                if "/" in det_obj:
+                    cur_uuid = url_to_uuid(det_obj)
+                else:
+                    cur_uuid = UUID(det_obj)
+
+                self.channel_objs[det_idx] = cur_uuid
 
     def __compute_bore2spin_quat__(self):
         quat = np.array(quat_rotation_z(self.boresight_rotangle_rad))