+++ litebird_sim/mapmaking/common.py
@@ -90,6 +90,8 @@ def get_map_making_weights(obs: Observation, check: bool) -> npt.NDArray:
     """
 
     try:
+        if isinstance(obs.net_ukrts, (float, int)):
+            obs.net_ukrts = obs.net_ukrts * np.ones(obs.n_detectors)
         weights = obs.sampling_rate_hz * obs.net_ukrts**2
     except AttributeError:
         weights = np.ones(obs.n_detectors)