+++ test/test_madam.py
@@ -6,9 +6,8 @@ from astropy.io import fits
 import numpy as np
 import astropy.units as u
 
-import litebird_sim
 import litebird_sim as lbs
-from litebird_sim.madam import _sort_obs_per_det, _ObsInMpiProcess
+from litebird_sim.madam import _sort_obs_per_det
 from litebird_sim.simulations import (
     MpiDistributionDescr,
     MpiProcessDescr,
@@ -35,6 +34,7 @@ def test_sort_obs_per_det():
         mpi_processes=[
             MpiProcessDescr(
                 mpi_rank=0,
+                numba_num_of_threads=1,
                 observations=[
                     MpiObservationDescr(
                         det_names=["A"],
@@ -62,6 +62,7 @@ def test_sort_obs_per_det():
             ),
             MpiProcessDescr(
                 mpi_rank=1,
+                numba_num_of_threads=1,
                 observations=[
                     MpiObservationDescr(
                         det_names=["A"],